refactor: split task record from queue
